chore: move edit page component
