+++ src/components/commons/Modals/InviteModal/InviteModal.module.scss
@@ -8,6 +8,10 @@
   padding: 3.2rem 2.8rem;
   border-radius: 0.8rem;
   z-index: $modal;
+
+  @include responsive("M") {
+    width: 90%;
+  }
 }
 
 .title {
