+++ src/pages/dashboard/[dashboardid]/edit/index.tsx
@@ -1,7 +1,7 @@
 import styles from "./edit.module.scss";
 import classNames from "classnames/bind";
-import DashboradEditTitleBox from "@/components/domains/edit/article/DashboradEditTitleBox";
-import DashboradEditMemberBox from "@/components/domains/edit/article/DashboradEditMemberBox";
+import DashboradEditTitleBox from "@/components/domains/edit/DashboradEditTitleBox";
+import DashboradEditMemberBox from "@/components/domains/edit/DashboradEditMemberBox";
 import Image from "next/image";
 import ResponseBtn from "@/components/commons/Buttons/ResponseButton";
 import BaseContainer from "@/components/commons/BaseContainer/BaseContainer";
